+++ C++/p011.cpp
@@ -31,7 +31,7 @@ What is the greatest product of four adjacent numbers in the same direction (up,
 
 #include <cmath>
 
-int p011()
+long long p011()
 {
     int a[20][20] = {{ 8,  2, 22, 97, 38, 15,  0, 40,  0, 75,  4, 05,  7, 78, 52, 12, 50, 77, 91,  8 },
         {49, 49, 99, 40, 17, 81, 18, 57, 60, 87, 17, 40, 98, 43, 69, 48,  4, 56, 62,  0},
